TheImperialGod/requirements.txt
aiohttp==3.6.3
async-timeout==3.0.1
asyncio==3.4.3
attrs==20.3.0
beautifulsoup4==4.9.3
captcha==0.3
certifi==2020.11.8
chardet==3.0.4
discord.py==1.5.1
ffmpeg==1.4
idna==2.10
multidict==4.7.6
Pillow==8.0.1
praw==7.1.0
prawcore==1.5.0
requests==2.24.0
six==1.15.0
soupsieve==2.0.1
update-checker==0.18.0
urllib3==1.25.11
websocket-client==0.57.0
wikipedia==1.4.0
yarl==1.5.1
youtube-dl==2020.11.1.1